+++ src/core/variable.jl
@@ -2,7 +2,7 @@
 function JuMP.value(x::Real) return x end
 
 "variable: `v[i]` for `i` in `bus`es"
-function variable_bus_voltage_magnitude_on_off(pm::_PM.AbstractPowerModel; nw::Int=pm.cnw, report::Bool=true)
+function variable_bus_voltage_magnitude_on_off(pm::_PM.AbstractPowerModel; nw::Int=nw_id_default, report::Bool=true)
     vm = _PM.var(pm, nw)[:vm] = JuMP.@variable(pm.model,
         [i in _PM.ids(pm, nw, :bus)], base_name="$(nw)_vm",
         lower_bound = 0.0,
@@ -14,7 +14,7 @@ function variable_bus_voltage_magnitude_on_off(pm::_PM.AbstractPowerModel; nw::I
 end
 
 "variable: `v[i]` for `i` in `bus`es"
-function variable_bus_voltage_magnitude_violation(pm::_PM.AbstractPowerModel; nw::Int=pm.cnw, report::Bool=true)
+function variable_bus_voltage_magnitude_violation(pm::_PM.AbstractPowerModel; nw::Int=nw_id_default, report::Bool=true)
     vm_vio = _PM.var(pm, nw)[:vm_vio] = JuMP.@variable(pm.model,
         [i in _PM.ids(pm, nw, :bus)], base_name="$(nw)_vm_vio",
         lower_bound = 0.0,
@@ -27,7 +27,7 @@ end
 
 
 "variable: `0 <= damage_gen[l] <= 1` for `l` in `gen`es"
-function variable_gen_damage_indicator(pm::_PM.AbstractPowerModel; nw::Int=pm.cnw, relax::Bool=false, report::Bool=true)
+function variable_gen_damage_indicator(pm::_PM.AbstractPowerModel; nw::Int=nw_id_default, relax::Bool=false, report::Bool=true)
     if relax == false
         z_gen_vars = JuMP.@variable(pm.model,
             [l in _PM.ids(pm, nw, :gen_damage)],
@@ -59,7 +59,7 @@ end
 
 
 "variable: `pg[j]` for `j` in `gen`"
-function variable_gen_power_real_damage(pm::_PM.AbstractPowerModel; nw::Int=pm.cnw, bounded::Bool=true, report::Bool=true)
+function variable_gen_power_real_damage(pm::_PM.AbstractPowerModel; nw::Int=nw_id_default, bounded::Bool=true, report::Bool=true)
     if bounded
         pg = _PM.var(pm, nw)[:pg] = JuMP.@variable(pm.model,
             [i in _PM.ids(pm, nw, :gen)], base_name="$(nw)_pg_dmg",
@@ -85,7 +85,7 @@ end
 
 
 "variable: `qq[j]` for `j` in `gen`"
-function variable_gen_power_imaginary_damage(pm::_PM.AbstractPowerModel; nw::Int=pm.cnw, bounded::Bool=true, report::Bool=true)
+function variable_gen_power_imaginary_damage(pm::_PM.AbstractPowerModel; nw::Int=nw_id_default, bounded::Bool=true, report::Bool=true)
     if bounded
         qg = _PM.var(pm, nw)[:qg] = JuMP.@variable(pm.model,
             [i in _PM.ids(pm, nw, :gen)], base_name="$(nw)_qg_dmg",
@@ -112,7 +112,7 @@ end
 
 
 "variable: `0 <= damage_branch[l] <= 1` for `l` in `branch`es"
-function variable_branch_damage_indicator(pm::_PM.AbstractPowerModel; nw::Int=pm.cnw, relax::Bool=false, report::Bool=true)
+function variable_branch_damage_indicator(pm::_PM.AbstractPowerModel; nw::Int=nw_id_default, relax::Bool=false, report::Bool=true)
     if relax == false
         z_branch_vars = JuMP.@variable(pm.model,
             [l in _PM.ids(pm, nw, :branch_damage)],
@@ -138,7 +138,7 @@ end
 
 
 "variable: `0 <= damage_storage[l] <= 1` for `l` in `storage`es"
-function variable_storage_damage_indicator(pm::_PM.AbstractPowerModel; nw::Int=pm.cnw, relax::Bool=false, report::Bool=true)
+function variable_storage_damage_indicator(pm::_PM.AbstractPowerModel; nw::Int=nw_id_default, relax::Bool=false, report::Bool=true)
     if relax == false
         z_storage_vars = JuMP.@variable(pm.model,
             [l in _PM.ids(pm, nw, :storage_damage)],
@@ -177,11 +177,11 @@ end
 
 
 "do nothing by default but some formulations require this"
-function variable_storage_current_damage(pm::_PM.AbstractPowerModel; nw::Int=pm.cnw, report::Bool=true)
+function variable_storage_current_damage(pm::_PM.AbstractPowerModel; nw::Int=nw_id_default, report::Bool=true)
 end
 
 ""
-function variable_storage_power_real_damage(pm::_PM.AbstractPowerModel; nw::Int=pm.cnw, report::Bool=true)
+function variable_storage_power_real_damage(pm::_PM.AbstractPowerModel; nw::Int=nw_id_default, report::Bool=true)
     inj_lb, inj_ub = _PM.ref_calc_storage_injection_bounds(_PM.ref(pm, nw, :storage), _PM.ref(pm, nw, :bus))
 
     ps = _PM.var(pm, nw)[:ps] = JuMP.@variable(pm.model,
@@ -203,7 +203,7 @@ end
 
 
 ""
-function variable_storage_power_imaginary_damage(pm::_PM.AbstractPowerModel; nw::Int=pm.cnw, report::Bool=true)
+function variable_storage_power_imaginary_damage(pm::_PM.AbstractPowerModel; nw::Int=nw_id_default, report::Bool=true)
     inj_lb, inj_ub = _PM.ref_calc_storage_injection_bounds(_PM.ref(pm, nw, :storage), _PM.ref(pm, nw, :bus))
 
     qs = _PM.var(pm, nw)[:qs] = JuMP.@variable(pm.model,
@@ -223,7 +223,7 @@ function variable_storage_power_imaginary_damage(pm::_PM.AbstractPowerModel; nw:
 end
 
 ""
-function variable_storage_power_control_imaginary_damage(pm::_PM.AbstractPowerModel; nw::Int=pm.cnw, report::Bool=true)
+function variable_storage_power_control_imaginary_damage(pm::_PM.AbstractPowerModel; nw::Int=nw_id_default, report::Bool=true)
     inj_lb, inj_ub = _PM.ref_calc_storage_injection_bounds(_PM.ref(pm, nw, :storage), _PM.ref(pm, nw, :bus))
 
     qsc = _PM.var(pm, nw)[:qsc] = JuMP.@variable(pm.model,
@@ -245,7 +245,7 @@ end
 
 
 "variable: `0 <= damage_bus[l] <= 1` for `l` in `bus`es"
-function variable_bus_damage_indicator(pm::_PM.AbstractPowerModel; nw::Int=pm.cnw, relax::Bool=false, report::Bool=true)
+function variable_bus_damage_indicator(pm::_PM.AbstractPowerModel; nw::Int=nw_id_default, relax::Bool=false, report::Bool=true)
     if relax == false
         z_bus_vars = JuMP.@variable(pm.model,
             [l in _PM.ids(pm, nw, :bus_damage)],
@@ -270,7 +270,7 @@ function variable_bus_damage_indicator(pm::_PM.AbstractPowerModel; nw::Int=pm.cn
 end
 
 
-function variable_bus_voltage_indicator(pm::_PM.AbstractPowerModel; nw::Int=pm.cnw, relax::Bool=false, report::Bool=true)
+function variable_bus_voltage_indicator(pm::_PM.AbstractPowerModel; nw::Int=nw_id_default, relax::Bool=false, report::Bool=true)
     if !relax
         z_voltage = _PM.var(pm, nw)[:z_voltage] = JuMP.@variable(pm.model,
             [i in _PM.ids(pm, nw, :bus)], base_name="$(nw)_z_voltage",
@@ -291,7 +291,7 @@ end
 
 
 ""
-function variable_bus_voltage_magnitude_sqr_on_off(pm::_PM.AbstractPowerModel; nw::Int=pm.cnw, report::Bool=true)
+function variable_bus_voltage_magnitude_sqr_on_off(pm::_PM.AbstractPowerModel; nw::Int=nw_id_default, report::Bool=true)
     w = _PM.var(pm, nw)[:w] = JuMP.@variable(pm.model,
         [i in _PM.ids(pm, nw, :bus)], base_name="$(nw)_w",
         lower_bound = 0,